disable container capability check on windows
