+++ pkg/agent/run_windows.go
@@ -1,0 +1,5 @@
+package agent
+
+func validate() error {
+	return nil
+}
